updating math
